Simplify getting started
* Rewrite readme with basic, consistent examples.
* Remove old examples.
* Move Go glue code into separate C source file.
* Build Go module entirely using MinGW-w64 on Windows.
* Adjustments to C++ code to get rid of warnings when compiling with MinGW-w64.
* Update Go test to call SetHtml instead of Navigate.
