+++ examples/go/basic.go
@@ -1,0 +1,12 @@
+package main
+
+import "github.com/webview/webview"
+
+func main() {
+	w := webview.New(false)
+	defer w.Destroy()
+	w.SetTitle("Basic Example")
+	w.SetSize(480, 320, webview.HintNone)
+	w.SetHtml("Thanks for using webview!")
+	w.Run()
+}
