+++ webview.cc
@@ -1,3 +1,1 @@
-// +build !windows
-
 #include "webview.h"
