+++ webview.h
@@ -836,12 +836,14 @@ using browser_engine = detail::cocoa_wkwebview_engine;
 #include <stdlib.h>
 #include <windows.h>
 
-#include "webview2.h"
+#include "WebView2.h"
 
+#ifdef _MSC_VER
 #pragma comment(lib, "ole32.lib")
 #pragma comment(lib, "shell32.lib")
 #pragma comment(lib, "shlwapi.lib")
 #pragma comment(lib, "user32.lib")
+#endif
 
 namespace webview {
 namespace detail {
